backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
pydantic-settings>=2.6
python-dotenv>=1.0
python-multipart>=0.0.12
httpx>=0.27
requests>=2.32
groq>=0.13
qdrant-client>=1.12
huggingface-hub>=0.26
lingua-language-detector>=2.0
python-docx>=1.1
pypdf>=5.1
gTTS>=2.5
sqlalchemy>=2.0
aiosqlite>=0.20
